// src/store/types/actionType.ts
export enum ActionType {
  ADD_PRODUCT = 'ADD_PRODUCT',
  GET_PRODUCT = 'GET_PRODUCT',
  DELETE_PRODUCT = 'DELETE_PRODUCT',
  UPDATE_PRODUCT = 'UPDATE_PRODUCT',
}
export enum ErrorType {
  FAILED_ACTION = 'FAILED_ACTION',
}
export enum ToastType {
  TOAST_SUCCESS_ADD = 'TOAST_SUCCESS_ADD',
  TOAST_SUCCESS_UPDATE = 'TOAST_SUCCESS_UPDATE',
  TOAST_SUCCESS_DELETE = 'TOAST_SUCCESS_DELETE',
  TOAST_FAILED = 'TOAST_FAILED',
  TOAST_NONE = 'TOAST_NONE',
}